pyproject.toml
[tool.poetry]
name = "kyokubot"
version = "0.1.100"
description = ""
authors = ["Kyoku <feikongtl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
nonebot2 = {extras = ["fastapi"], version = "^2.3.2"}
nonebot-adapter-onebot = "^2.4.4"
nb-cli = "^1.4.1"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
pymysql = "^1.1.1"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "kyokubot.start:main"

[kyokubot.config]
# onebot连接部分
host = "0.0.0.0"
port = 23333
# 数据库设置部分，数据库可用postgresql, mysql, sqlite三种类型
# 请确保目标主机中有名为kyokubotdb的数据库且提供的用户名对其具有访问权限
# 如使用sqlite数据库那么其他配置项不起实际作用
db_type = "postgresql"
db_host= "localhost"
db_port = 5432
db_username = "kyokubot"
db_passwd = "kyokubot"